+++ packages/core-react-ui/app/app.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Button } from 'core-react-ui';
 
 import * as styleRefs from './app.treat';
 import { useStyles } from 'react-treat';
@@ -8,11 +7,5 @@ import { ButtonSize } from 'core-design-system';
 export function App() {
   const styles = useStyles(styleRefs);
 
-  return (
-    <div>
-      <Button size={ButtonSize.large}>hey</Button>
-      <Button size={ButtonSize.medium}>hey</Button>
-      <Button size={ButtonSize.small}>hey</Button>
-    </div>
-  );
+  return <div>hello</div>;
 }