+++ packages/core-react-ui/src/components/seo/metadata-website.tsx
@@ -13,6 +13,7 @@ export const MetadataOpenGraphWebsiteTags: React.FC<Props> = (props) => {
 };
 
 export function getMetadataOpenGraphWebsiteTags({
+  lang,
   locale,
   title,
   url,
@@ -23,9 +24,10 @@ export function getMetadataOpenGraphWebsiteTags({
   localeAlternate,
 }: Props) {
   return [
-    <html lang={locale ? locale : 'en_US'} prefix="og: https://ogp.me/ns#" />,
+    <html lang={lang ? lang : 'en'} prefix="og: https://ogp.me/ns#" />,
     title ? <title>{title}</title> : null,
     url ? <link rel="canonical" href={url} /> : null,
+    url ? <meta property="og:url" content={url} /> : null,
     description ? <meta name="description" content={description} /> : null,
     title ? <meta property="og:title" content={title} /> : null,
     type ? <meta property="og:type" content={type} /> : null,