+++ packages/core-react-ui/src/index.tsx
@@ -1,6 +1,3 @@
-export * from './components/button/button';
-export * from './components/button/button.treat';
-
 export * from './components/seo/meta-basic';
 export * from './components/seo/metadata-website';
 export * from './components/seo/metadata-twitter';