+++ packages/mir-website/src/components/action-section/action-section.tsx
@@ -31,7 +31,7 @@ export const ActionSection: React.FC<OwnProps & AllHTMLAttributes<any>> = (props
         <div className={`${styles.buttonContainer}`}>
           <Button
             className={`${styles.button}`}
-            type={props.variant === 'Default' ? 'primaryDefault' : 'primaryReversed'}
+            variantType={props.variant === 'Default' ? 'primaryDefault' : 'primaryReversed'}
             size="medium"
             variant="text"
           >