+++ packages/mir-website/src/components/info-section/info-section-type-2.tsx
@@ -42,7 +42,7 @@ export const InfoSectionType2: React.FC<OwnProps> = (props) => {
           {props.text}
         </Paragraph>
 
-        <Button type={'primaryDefault'} variant={'text'} size={'medium'} className={styles.button}>
+        <Button variantType={'primaryDefault'} variant={'text'} size={'medium'} className={styles.button}>
           {props.actionText}
         </Button>
       </div>