+++ packages/mir-website/src/components/input.tsx
@@ -7,10 +7,10 @@ interface OwnProps {
   placeholder: string;
 }
 
-export const Input: React.FC<OwnProps & HTMLAttributes<any>> = (props) => {
+export const Input: React.FC<OwnProps & HTMLAttributes<any>> = ({ className, variant, placeholder, ...props }) => {
   const styles = useStyles(styleRefs);
 
   return (
-    <input className={`${styles.wrapper} ${styles[props.variant]}`} type="text" placeholder={`${props.placeholder}`} />
+    <input className={`${styles.wrapper} ${styles[variant]}`} type="text" placeholder={`${placeholder}`} {...props} />
   );
 };