+++ packages/mir-website/src/graphql/testimonial-section.fragment.ts
@@ -1,0 +1,14 @@
+import { graphql } from 'gatsby';
+
+export const testimonialSectionQuery = graphql`
+  fragment TestimonialSection on GraphCMS_TestimonialSection {
+    title
+    titleHighlight
+    description
+    testomonials {
+      message
+      authorName
+      authorTitle
+    }
+  }
+`;