+++ packages/mir-website/src/layouts/en.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Layout } from './index';
-
-export const LayoutEN: React.FC<{}> = (props) => <Layout {...props} />;
-
-export default LayoutEN;