+++ packages/mir-website/src/pages/404.fr.tsx
@@ -6,62 +6,68 @@ import {
 } from 'core-react-ui';
 import { graphql } from 'gatsby';
 import React from 'react';
-import { Helmet } from 'react-helmet-async';
+import { Helmet } from 'react-helmet';
 import { NotFoundPageFrQuery } from '../../types/graphql-types';
-import Layout from '../layouts';
+import { Layout } from '../layouts';
+import { NotFoundPageTemplate } from '../templates/not-found-page.template';
 
 export const query = graphql`
   query NotFoundPageFR {
     site {
       ...SiteMetadata
     }
-    bannerImageMobile: file(name: { eq: "ImageOffice05" }) {
-      ...MobileFluidImage
-    }
-    bannerImageDesktop: file(name: { eq: "ImageOffice05" }) {
-      ...DesktopFluidImage
-    }
     gcms {
-      notFoundPages(first: 1, locales: fr) {
-        ...NotFoundPage
+      companyInfos(first: 1) {
+        ...CompanyInfo
       }
-      metadataWebsites(first: 1) {
-        siteName
-        siteUrl
+      pages(where: { name: "Not Found" }, locales: [fr, en]) {
+        ...Page
       }
-      metadataTwitters(first: 1) {
-        creator
-        site
+      routes: pages(where: { NOT: { name: "Not Found" } }, locales: [fr, en]) {
+        name
+        title
+        route
       }
     }
   }
 `;
 
-export interface PageProps {
+interface PageProps {
   data: NotFoundPageFrQuery;
   location: Location;
 }
 
 const NotFoundPage: React.FC<PageProps> = (props) => {
   return (
-    <Layout>
+    <Layout
+      location={props.location}
+      logoURL={props.data.gcms.companyInfos[0].logo?.url}
+      linkedinPageURL={props.data.gcms.companyInfos[0].linkedinPageUrl}
+      facebookPageURL={props.data.gcms.companyInfos[0].facebookPageUrl}
+      instagramPageURL={props.data.gcms.companyInfos[0].instagramPageUrl}
+      twitterPageURL={props.data.gcms.companyInfos[0].twitterPageUrl}
+      pages={props.data.gcms.routes}
+    >
       <Helmet>
-        <title>test</title>
         {getMetaBasicTags()}
         {getMetadataOpenGraphWebsiteTags({
           type: OPEN_GRAPH_TYPE.WEBSITE,
-          title: 'test',
-          url: 'test',
-          description: 'test',
-          site_name: 'test',
-          locale: 'test',
-          localeAlternate: 'test',
+          title: `${props.data.gcms.pages[0]?.title}`,
+          url: `${props.data.site?.siteMetadata?.siteUrl}${props.data.gcms.pages[0]?.route}`,
+          description: `${props.data.gcms.pages[0]?.description}`,
+          image: `${props.data.gcms.pages[0]?.bannerImages[0]?.url}`,
+          site_name: `${props.data.gcms.companyInfos[0].metadataSiteName}`,
+          lang: 'fr',
+          locale: 'fr_CA',
+          localeAlternate: 'en_CA',
         })}
         {getMetadataTwitterTags({
           card: 'summary',
+          creator: `${props.data.gcms.companyInfos[0].metadataTwitterCreator}`,
+          site: `${props.data.gcms.companyInfos[0].metadataTwitter}`,
         })}
       </Helmet>
-      <div>Not Found</div>
+      <NotFoundPageTemplate {...props} />
     </Layout>
   );
 };