+++ packages/mir-website/src/templates/candidates-page.template.tsx
@@ -2,40 +2,47 @@ import React from 'react';
 import { useStyles } from 'react-treat';
 import * as stylesRef from '../templates/candidates-page.treat';
 import { BannerSecondary } from '../components/banner-secondary';
-import { Footer } from '../components/footer';
-import { NavBar } from '../components/nav-bar';
-import { Newsletter } from '../components/newsletter/newsletter';
+
 import { InfoSectionType5 } from '../components/info-section/info-section-type-5';
 import { InfoSectionType2 } from '../components/info-section/info-section-type-2';
 import { ActionSection } from '../components/action-section/action-section';
 import { JobSection } from '../components/job-section/job-section';
 import { Process } from '../components/process-section/process-section';
-import { CandidatesPageProps } from '../pages/candidates.en';
+import { CandidatePageFrQuery } from '../../types/graphql-types';
+import { theme } from '../design-system';
 
-export const Candidate: React.FC<CandidatesPageProps> = ({ data, location }) => {
+interface PageProps {
+  data: CandidatePageFrQuery;
+  location: Location;
+}
+
+export const Candidate: React.FC<PageProps> = ({ data, location }) => {
   const styles = useStyles(stylesRef);
 
-  const actionSection1 = data.gcms.candidatePages[0].actionSections[0];
-  const actionSection2 = data.gcms.candidatePages[0].actionSections[1];
-  const section1 = data.gcms?.candidatePages[0]?.infoSections[0];
-  const section2 = data.gcms?.candidatePages[0]?.infoSections[1];
-  const jobSection = data.gcms?.candidatePages[0]?.jobSection;
-  const processSection = data.gcms?.candidatePages[0]?.processSection;
+  const actionSection1 = data.gcms.pages[0].actionSections[0];
+  const actionSection2 = data.gcms.pages[0].actionSections[1];
+  const section1 = data.gcms?.pages[0]?.infoSections[0];
+  const section2 = data.gcms?.pages[0]?.infoSections[1];
+  const jobSection = data.gcms?.pages[0]?.jobSections[0];
+  const processSection = data.gcms?.pages[0]?.processSections[0];
+
+  const sources = [
+    data?.bannerImageMobile?.childImageSharp?.fluid,
+    {
+      ...data?.bannerImageDesktop?.childImageSharp?.fluid,
+      media: `(min-width: ${theme.layout.breakpoints.desktopSmall.px})`,
+    },
+  ];
 
   return (
-    <div className={`${styles.wrapper}`}>
-      <NavBar></NavBar>
-      <BannerSecondary
-        imageData={data.bannerImage?.childImageSharp?.fluid}
-        title={data?.gcms?.candidatePages[0]?.bannerTitle}
-      ></BannerSecondary>
+    <main className={`${styles.wrapper}`}>
+      <BannerSecondary imageData={sources} title={data?.gcms?.pages[0]?.bannerTitle}></BannerSecondary>
       {actionSection1 ? <ActionSection variant={'Default'} /> : null}
       {section1 && section1.type === 'type5' ? <InfoSectionType5 align="AlignContentLeft" {...section1} /> : null}
       {jobSection ? <JobSection jobSection={jobSection} /> : null}
       {section2 && section2.type === 'type2' ? <InfoSectionType2 align="AlignContentRight" {...section2} /> : null}
       {processSection ? <Process processSection={processSection} /> : null}
       {actionSection2 ? <ActionSection variant={'reversed'} /> : null}
-      <Footer></Footer>
-    </div>
+    </main>
   );
 };